Adding new routes

+++ controllers/users.controller.js
@@ -13,6 +13,13 @@ const getUsers = async (req, res, next) => {
 const createUser = async (req, res, next) => {
   try {
     const { first_name, last_name, email, role } = req.body;
+    const userReq=req.user;
+
+    if(userReq.role!=='Librarian'){
+      return res.status(401).json({
+        message: "Unauthorized",
+      });
+    }
 
     //creating a default password
     const password = "password";
@@ -44,6 +51,13 @@ const updateUser = async (req, res, next) => {
 
 const deleteUser = async (req, res, next) => {
   try {
+    const userReq=req.user;
+
+    if(userReq.role!=='Librarian'){
+      return res.status(401).json({
+        message: "Unauthorized",
+      });
+    }
     const user = await pool.query("DELETE FROM users WHERE id=$1 RETURNING *", [
       req.params.id,
     ]);
@@ -64,7 +78,34 @@ const getUserById = async (req, res, next) => {
   }
 };
 
+const changePassword = async (req, res, next) => {
+  try {
+    const { oldPassword, newPassword } = req.body;
+    const userCtx = req.user; 
+    const user = await pool.query(
+      "SELECT * FROM users WHERE id=$1",
+      [userCtx.id]
+    );
+    const isMatch = await bcrypt.compare(oldPassword, user.rows[0].password);
+    if (!isMatch) {
+      return res.status(401).json({
+        message: "Unauthorized",
+      });
+    }
+    const salt = await bcrypt.genSalt(10);
+    const hashedPassword = await bcrypt.hash(newPassword, salt);
+    const userUpdate = await pool.query(
+      "UPDATE users SET password=$1 WHERE id=$2 RETURNING *",
+      [hashedPassword, userCtx.id]
+    );
+    res.status(200).json(userUpdate.rows[0]);
+  } catch (error) {
+    next(error);
+  }
+}
+
 module.exports = {
+  changePassword,
   createUser,
   deleteUser,
   getUserById,
